Restructure ConfirmPage child components; change icon for bytes
