+++ src/app/templates/ConnectBanner.tsx
